+++ src/main/java/dev/willyelton/crystal_tools/datagen/CrystalToolsBlockTags.java
@@ -2,9 +2,11 @@ package dev.willyelton.crystal_tools.datagen;
 
 import dev.willyelton.crystal_tools.CrystalTools;
 import dev.willyelton.crystal_tools.Registration;
+import dev.willyelton.crystal_tools.common.tags.CrystalToolsTags;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.world.level.block.Blocks;
 import net.neoforged.neoforge.common.data.BlockTagsProvider;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
@@ -24,5 +26,8 @@ public class CrystalToolsBlockTags extends BlockTagsProvider {
                 Registration.CRYSTAL_FURNACE.get(),
                 Registration.CRYSTAL_GENERATOR.get(),
                 Registration.CRYSTAL_QUARRY.get());
+
+        tag(CrystalToolsTags.AUTO_OUTPUT_BLACKLIST).add(
+                Blocks.HOPPER);
     }
 }